+++ src/Common/AssemblyInfo.fs
@@ -4,9 +4,9 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FSharp.Formatting")>]
 [<assembly: AssemblyProductAttribute("FSharp.Formatting")>]
 [<assembly: AssemblyDescriptionAttribute("A package for building great F# documentation, samples and blogs")>]
-[<assembly: AssemblyVersionAttribute("2.2.4")>]
-[<assembly: AssemblyFileVersionAttribute("2.2.4")>]
+[<assembly: AssemblyVersionAttribute("2.2.6")>]
+[<assembly: AssemblyFileVersionAttribute("2.2.6")>]
 ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "2.2.4"
+    let [<Literal>] Version = "2.2.6"