+++ src/FSharp.Literate/ParseMarkdown.fs
@@ -5,6 +5,6 @@ open FSharp.Markdown
 module ParseMarkdown =
   /// Parse the specified Markdown document and return it
   /// as `LiterateDocument` (without processing code snippets)
-  let parseMarkdown text =
+  let parseMarkdown file text =
     let doc = Markdown.Parse(text)
-    LiterateDocument(doc.Paragraphs, "", doc.DefinedLinks, LiterateSource.Markdown text, [])
+    LiterateDocument(doc.Paragraphs, "", doc.DefinedLinks, LiterateSource.Markdown text, file, [])